// Frey/app/clients/proto/frey.pb.ts
export interface AuthRequestRegistrationPayload {
  email: string;
  title: string;
  password: string;
  sendEmail: boolean;
}

export interface AuthRequestRegistrationResponse {
  approvalCode: string;
}

export interface AuthApproveRegistrationPayload {
  approvalCode: string;
}

export interface AuthApproveRegistrationResponse {
  personId: number;
}

export interface AuthAuthorizePayload {
  email: string;
  password: string;
}

export interface AuthAuthorizeResponse {
  personId: number;
  authToken: string;
}

export interface AuthQuickAuthorizePayload {
  personId: number;
  authToken: string;
}

export interface AuthQuickAuthorizeResponse {
  authSuccess: boolean;
}

export interface AuthMePayload {
  personId: number;
  authToken: string;
}

export interface AuthMeResponse {
  personId: number;
  country: string;
  city: string;
  email: string;
  phone: string;
  tenhouId: string;
  title: string;
  groups: number[];
  hasAvatar: boolean;
  lastUpdate: string;
}

export interface AuthChangePasswordPayload {
  email: string;
  password: string;
  newPassword: string;
}

export interface AuthChangePasswordResponse {
  authToken: string;
}

export interface AuthRequestResetPasswordPayload {
  email: string;
  sendEmail: boolean;
}

export interface AuthRequestResetPasswordResponse {
  resetToken: string;
}

export interface AuthApproveResetPasswordPayload {
  email: string;
  resetToken: string;
}

export interface AuthApproveResetPasswordResponse {
  newTmpPassword: string;
}
// ...
export const AuthApproveRegistrationPayload = {
  /**
   * Serializes AuthApproveRegistrationPayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<AuthApproveRegistrationPayload>,
  ): Uint8Array {
    return AuthApproveRegistrationPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AuthApproveRegistrationPayload from protobuf.
   */
  decode: function (bytes: ByteSource): AuthApproveRegistrationPayload {
    return AuthApproveRegistrationPayload._readMessage(
      AuthApproveRegistrationPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AuthApproveRegistrationPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AuthApproveRegistrationPayload>,
  ): AuthApproveRegistrationPayload {
    return {
      approvalCode: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AuthApproveRegistrationPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.approvalCode) {
      writer.writeString(1, msg.approvalCode);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AuthApproveRegistrationPayload,
    reader: protoscript.BinaryReader,
  ): AuthApproveRegistrationPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.approvalCode = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const AuthApproveRegistrationResponse = {
  /**
   * Serializes AuthApproveRegistrationResponse to protobuf.
   */
  encode: function (
    msg: PartialDeep<AuthApproveRegistrationResponse>,
  ): Uint8Array {
    return AuthApproveRegistrationResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AuthApproveRegistrationResponse from protobuf.
   */
  decode: function (bytes: ByteSource): AuthApproveRegistrationResponse {
    return AuthApproveRegistrationResponse._readMessage(
      AuthApproveRegistrationResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AuthApproveRegistrationResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AuthApproveRegistrationResponse>,
  ): AuthApproveRegistrationResponse {
    return {
      personId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AuthApproveRegistrationResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.personId) {
      writer.writeInt32(1, msg.personId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AuthApproveRegistrationResponse,
    reader: protoscript.BinaryReader,
  ): AuthApproveRegistrationResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.personId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const AuthAuthorizePayload = {
  /**
   * Serializes AuthAuthorizePayload to protobuf.
   */
  encode: function (msg: PartialDeep<AuthAuthorizePayload>): Uint8Array {
    return AuthAuthorizePayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AuthAuthorizePayload from protobuf.
   */
  decode: function (bytes: ByteSource): AuthAuthorizePayload {
    return AuthAuthorizePayload._readMessage(
      AuthAuthorizePayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AuthAuthorizePayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AuthAuthorizePayload>,
  ): AuthAuthorizePayload {
    return {
      email: "",
      password: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AuthAuthorizePayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.email) {
      writer.writeString(1, msg.email);
    }
    if (msg.password) {
      writer.writeString(2, msg.password);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AuthAuthorizePayload,
    reader: protoscript.BinaryReader,
  ): AuthAuthorizePayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.email = reader.readString();
          break;
        }
        case 2: {
          msg.password = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const AuthAuthorizeResponse = {
  /**
   * Serializes AuthAuthorizeResponse to protobuf.
   */
  encode: function (msg: PartialDeep<AuthAuthorizeResponse>): Uint8Array {
    return AuthAuthorizeResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AuthAuthorizeResponse from protobuf.
   */
  decode: function (bytes: ByteSource): AuthAuthorizeResponse {
    return AuthAuthorizeResponse._readMessage(
      AuthAuthorizeResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AuthAuthorizeResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AuthAuthorizeResponse>,
  ): AuthAuthorizeResponse {
    return {
      personId: 0,
      authToken: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AuthAuthorizeResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.personId) {
      writer.writeInt32(1, msg.personId);
    }
    if (msg.authToken) {
      writer.writeString(2, msg.authToken);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AuthAuthorizeResponse,
    reader: protoscript.BinaryReader,
  ): AuthAuthorizeResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.personId = reader.readInt32();
          break;
        }
        case 2: {
          msg.authToken = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};
// ...
export const AuthApproveResetPasswordPayload = {
  /**
   * Serializes AuthApproveResetPasswordPayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<AuthApproveResetPasswordPayload>,
  ): Uint8Array {
    return AuthApproveResetPasswordPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AuthApproveResetPasswordPayload from protobuf.
   */
  decode: function (bytes: ByteSource): AuthApproveResetPasswordPayload {
    return AuthApproveResetPasswordPayload._readMessage(
      AuthApproveResetPasswordPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AuthApproveResetPasswordPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AuthApproveResetPasswordPayload>,
  ): AuthApproveResetPasswordPayload {
    return {
      email: "",
      resetToken: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AuthApproveResetPasswordPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.email) {
      writer.writeString(1, msg.email);
    }
    if (msg.resetToken) {
      writer.writeString(2, msg.resetToken);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AuthApproveResetPasswordPayload,
    reader: protoscript.BinaryReader,
  ): AuthApproveResetPasswordPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.email = reader.readString();
          break;
        }
        case 2: {
          msg.resetToken = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const AuthApproveResetPasswordResponse = {
  /**
   * Serializes AuthApproveResetPasswordResponse to protobuf.
   */
  encode: function (
    msg: PartialDeep<AuthApproveResetPasswordResponse>,
  ): Uint8Array {
    return AuthApproveResetPasswordResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AuthApproveResetPasswordResponse from protobuf.
   */
  decode: function (bytes: ByteSource): AuthApproveResetPasswordResponse {
    return AuthApproveResetPasswordResponse._readMessage(
      AuthApproveResetPasswordResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AuthApproveResetPasswordResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AuthApproveResetPasswordResponse>,
  ): AuthApproveResetPasswordResponse {
    return {
      newTmpPassword: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AuthApproveResetPasswordResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.newTmpPassword) {
      writer.writeString(1, msg.newTmpPassword);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AuthApproveResetPasswordResponse,
    reader: protoscript.BinaryReader,
  ): AuthApproveResetPasswordResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.newTmpPassword = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};